Time: 7 ms (13.96%), Space: 43.9 MB (41.21%) - LeetHub

+++ 0201-bitwise-and-of-numbers-range/0201-bitwise-and-of-numbers-range.java
@@ -10,6 +10,6 @@ class Solution {
         for(int i = 30;i>=diffBits;i--){
             mask+=(long)Math.pow(2,i);
         }
-        return (int)(left & right & mask);
+        return left & right & mask;
     }
 }
